make more optimized custom deletion query
Also Scheduling for User is moved in a separate class
