+++ Task2/src/main/java/org/neyaank/task2/user/UserRepository.java
@@ -6,6 +6,8 @@
 package org.neyaank.task2.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
@@ -16,6 +18,11 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findById(int id);
     Optional<User> findByVerificationCode(String verificationCode);
-    List<User> findAllByVerificationStartTimeLessThanEqualAndVerifiedFalse(LocalDateTime startTime);
     Optional<User> findByEmail(String email);
+    @Modifying
+    @Query("DELETE FROM User " +
+            "WHERE verified = FALSE " +
+            "AND verificationStartTime <= :startTime")
+    int deleteUnverifiedAndOldEnough(LocalDateTime startTime);
+
 }
